Add files via upload
Design and Analysis of a Waste Water Treatment Plant in Chhalaila.
The water we use for showers, laundry, and sanitation doesn’t simply disappear
down the drain. Instead, it undergoes a multi-stage process to remove contaminants
and transform it into a safe and reusable resource.

This project delves into the critical world of wastewater treatment. We will explore
the methods of treating wastewater, from the removal of large solids to the breakdown of harmful bacteria. We did energy balance,material balance and equipment
design of the various equipment involved in the wastewater treatment plant. Along
with these calculations pertaining to the equipment, we also did the overall tentative
costing for the establishment of the wastewater treatment plant.

Through the application of fundamental principles such as material balances, energy balances, and equipment design, we aim to develop an innovative and efficient
treatment model tailored to our village’s specific requirements.